use pre-submit form to check for all sections being completed in both apply and review steps
