+++ app/form_models/jobseekers/job_application/pre_submit_form.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+module Jobseekers
+  module JobApplication
+    class PreSubmitForm
+      include ActiveModel::Model
+
+      attr_accessor :completed_steps, :all_steps
+
+      validate :all_steps_completed?
+
+      def all_steps_completed?
+        all_steps.each do |step|
+          next if step.in?(completed_steps)
+
+          errors.add(
+            :base,
+            # I18n.t("activemodel.errors.models.jobseekers/job_application/review_form.attributes.#{step}.incomplete"),
+            "#{step}.incomplete".to_sym,
+            )
+        end
+      end
+    end
+  end
+end
+
